pyproject.toml
[project]
name = "hybridrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "lightrag-hku>=1.4.8.2",
    "ollama>=0.6.0",
]
